Shot of the towers

+++ towerdefense-git/TowerDefense2/app/src/main/java/com/example/towerdefense/Buttons_tower_upgrade.java
@@ -19,7 +19,7 @@ public class Buttons_tower_upgrade extends Buttons {
     }
 
     public void getEvent(MotionEvent event, List<Towers> towers, Context context) {
-        towers.add(new Tower2(towers.get(towerIndex).getX(), towers.get(towerIndex).getY(), context));
+        towers.add(new Tower2((int) towers.get(towerIndex).getX(), (int) towers.get(towerIndex).getY(), context));
         towers.remove(towerIndex);
     }
